Add options to force headers height/width
These new options (forceHeadersWidth and
forceHeadersHeight) allows the plugin to work with
tables where the cells content are wider or higher
than their matching headers. It will essentially
force the cloned headers to use the same size as
the original ones, when needed.

That being said, it needs to loop on every rows or
columns (or both), depending on which option is
turned on, potentially being quite heavy to process
with huge tables.